+++ tests/dummy/mirage/serializers/application.js
@@ -1,0 +1,3 @@
+import ActiveModelSerializer from 'ember-cli-mirage/serializers/active-model-serializer';
+
+export default ActiveModelSerializer;